Extract the address module
