+++ address/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "address"
+version = "0.1.0"
+authors = ["The PolkaX Authors"]
+edition = "2018"
+
+[dependencies]
+#bls = { package = "bls-signatures", version = "0.3" }
+varint = { package = "unsigned-varint", version = "0.2.3" }
+blake2-rfc = { package = "blake2-rfc", version = "0.2.18" }
+libsecp256k1 = { version = "0.3.0" }
+derive_more = { version = "0.99.2"}
+data-encoding = { package = "data-encoding", version = "2.0.0" }
+hex = "0.4.0"
+serde_json = "1.0.41"
+serde_cbor = { git = "https://github.com/PolkaX/cbor"}
