+++ chain/src/types/message.rs
@@ -1,6 +1,7 @@
 // Copyright 2019 PolkaX Authors. Licensed under GPL-3.0.
 
-use crate::types::{Address, BigInt};
+use crate::types::BigInt;
+use address::Address;
 
 pub struct Message {
     to: Address,
