+++ wallet/Cargo.toml
@@ -6,16 +6,15 @@ edition = "2018"
 
 [dependencies]
 bls = { package = "bls-signatures", version = "0.3" }
-varint = { package = "unsigned-varint", version = "0.2.3" }
-blake2-rfc = { package = "blake2-rfc", version = "0.2.18" }
 libsecp256k1 = { version = "0.3.0" }
 derive_more = { version = "0.99.2"}
-data-encoding = { package = "data-encoding", version = "2.0.0" }
 hex = "0.4.0"
 rand = "0.7.2"
 serde_json = "1.0.41"
 parking_lot = "0.9.0"
 failure = "0.1.6"
+
+address = {path = "../address"}
 
 [dev-dependencies]
 tempfile = "3.1.0"
